:recycle: [fix] 파고 예측 정보 삭제, 강수량 예측 정보 추가
