+++ weatherAPI/services/weatherAPI.service.js
@@ -2,18 +2,13 @@ require('dotenv').config();
 const { CustomError } = require('../../utils/Error');
 const axios = require('axios');
 
-const {
-  today_func,
-  find_tide_observatory,
-  find_wave_height_observatory,
-} = require('../observatoryFunc/find_observatory');
+const { today_func, all_time_info_in_today, find_tide_observatory } = require('../observatoryFunc/find_observatory');
 
 // data에서 지금날짜 기준 2일 후의 data까지 찾아오는 함수: open weather api 전용
 const find_after2days_from_now_data = (weather_data, today, after2days_from_today) => {
   const now = weather_data.findIndex((data) => data.date >= today);
   const after2days_from_now = weather_data.findIndex((data) => data.date >= after2days_from_today);
   const result = weather_data.slice(now, after2days_from_now + 1);
-  // const result = weather_data.slice(now + 3, after2days_from_now + 4);
 
   return result.map((data) => {
     return {
@@ -21,6 +16,7 @@ const find_after2days_from_now_data = (weather_data, today, after2days_from_toda
       temp: data.temp,
       wind_speed: data.wind_speed,
       wind_deg: data.wind_deg,
+      rain: data.rain,
       date: data.date.split(' ')[1].split(':')[0],
       original_time: data.date,
     };
@@ -58,6 +54,7 @@ const get_weather = async (lat, lon) => {
           temp: parseInt(Math.round(data.main.temp) - 273.15),
           wind_speed: data.wind.speed,
           wind_deg: data.wind.deg,
+          rain: data.rain ? data.rain['3h'] : 0,
           date: data.dt_txt,
         };
       });
@@ -78,6 +75,7 @@ const get_weather = async (lat, lon) => {
   return data;
 };
 
+// 조석예보 api 요청 함수
 const axios_tide_info = async (url) => {
   const map_data = await axios
     .get(url)
@@ -114,15 +112,7 @@ const get_tide_info = async (lat, lon) => {
   const tide = [...today_tide.data, ...after2days_tide.data, ...after3days_tide.data];
 
   const index = tide.findIndex((data) => {
-    const date = new Date();
-    const year = date.getFullYear(); // 년
-    const month = ('0' + (date.getMonth() + 1)).slice(-2); // 월
-    const day = date.getDate(); // 일
-    const hour = date.getHours(); // 시간
-    const minutes = ('0' + date.getMinutes()).slice(-2); // 분
-    const seconds = ('0' + date.getSeconds()).slice(-2); // 초
-
-    const now = `${year}-${month}-${day} ${hour}:${minutes}:${seconds}`;
+    const now = all_time_info_in_today().YY_MM_DD_HH_MM_SS;
 
     return now < data.tph_time;
   });
@@ -130,34 +120,7 @@ const get_tide_info = async (lat, lon) => {
   return tide.slice(index, index + 8);
 };
 
-// 바다누리: 파고
-//! database에 저장 안하고 바로 불러와 처리하기
-//! 크롤링 단방향 요청 시에만 긁어오기
-const get_wave_height_info = async (lat, lon) => {
-  const obs_post_id = await find_wave_height_observatory(lat, lon);
-  const today = await today_func();
-
-  const open_api = 'http://www.khoa.go.kr/api/oceangrid/obsWaveHight/search.do?ServiceKey=';
-  const service_key = `${process.env.OPEN_BADANURI_API_KEYS}`;
-  const obs_code = `&ObsCode=${obs_post_id}`; // 관측소 번호
-  const date = `&Date=${today}`; // 오늘 날짜
-
-  const search_info = open_api + service_key + obs_code + date + `&ResultType=json`;
-
-  const data = axios
-    .get(search_info)
-    .then((res) => {
-      return res.data.result;
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-
-  return data;
-};
-
 module.exports = {
   get_weather,
   get_tide_info,
-  get_wave_height_info,
 };
