+++ weatherAPI/weatherAPI.controller.js
@@ -1,14 +1,12 @@
-const { get_weather, get_tide_info, get_wave_height_info } = require('./services/weatherAPI.service');
+const { get_weather, get_tide_info } = require('./services/weatherAPI.service');
 
 const get_weather_api = async (req, res) => {
   try {
     const { lat, lon, place_name } = req.query;
     // open weather: 풍향, 풍속, 기온
-    const open_weather = await get_weather(lat, lon); // console.log('open_weather: ', open_weather);
+    const open_weather = await get_weather(lat, lon);
     // 바다누리: 만조, 간조
     const tide = await get_tide_info(lat, lon);
-    // 바다누리: 파고
-    const wave_height = await get_wave_height_info(lat, lon);
 
     return res.status(200).send({
       message: '날씨 정보 검색 완료',
@@ -17,7 +15,6 @@ const get_weather_api = async (req, res) => {
       lat: lat,
       weather: open_weather,
       tide_info: tide,
-      wave_info: wave_height.data,
     });
   } catch (error) {
     console.log(error);
